+++ sdk/storage/azdatalake/internal/generated_blob/models.go
@@ -6,6 +6,12 @@
 
 package generated_blob
 
+import (
+	"encoding/xml"
+	"github.com/Azure/azure-sdk-for-go/sdk/azcore/to"
+	"net/url"
+)
+
 type TransactionalContentSetter interface {
 	SetCRC64([]byte)
 	SetMD5([]byte)
@@ -62,4 +68,65 @@ func (p *PageBlobClientUploadPagesFromURLOptions) SetSourceContentCRC64(v []byte
 
 func (p *PageBlobClientUploadPagesFromURLOptions) SetSourceContentMD5(v []byte) {
 	p.SourceContentMD5 = v
+}
+
+// Custom UnmarshalXML functions for types that need special handling.
+
+// UnmarshalXML implements the xml.Unmarshaller interface for type PathPrefix.
+func (p *PathPrefix) UnmarshalXML(dec *xml.Decoder, start xml.StartElement) error {
+	type alias PathPrefix
+	aux := &struct {
+		*alias
+		PathName *PathName `xml:"Name"`
+	}{
+		alias: (*alias)(p),
+	}
+	if err := dec.DecodeElement(aux, &start); err != nil {
+		return err
+	}
+	if aux.PathName != nil {
+		if aux.PathName.Encoded != nil && *aux.PathName.Encoded {
+			name, err := url.QueryUnescape(*aux.PathName.Content)
+
+			if err != nil {
+				return err
+			}
+			p.Name = to.Ptr(string(name))
+		} else {
+			p.Name = aux.PathName.Content
+		}
+	}
+	return nil
+}
+
+// UnmarshalXML implements the xml.Unmarshaller interface for type PathItem.
+func (p *PathItem) UnmarshalXML(dec *xml.Decoder, start xml.StartElement) error {
+	type alias PathItem
+	aux := &struct {
+		*alias
+		PathName   *PathName            `xml:"Name"`
+		Metadata   additionalProperties `xml:"Metadata"`
+		OrMetadata additionalProperties `xml:"OrMetadata"`
+	}{
+		alias: (*alias)(p),
+	}
+	if err := dec.DecodeElement(aux, &start); err != nil {
+		return err
+	}
+	p.Metadata = (map[string]*string)(aux.Metadata)
+	p.OrMetadata = (map[string]*string)(aux.OrMetadata)
+	if aux.PathName != nil {
+		if aux.PathName.Encoded != nil && *aux.PathName.Encoded {
+			name, err := url.QueryUnescape(*aux.PathName.Content)
+
+			// name, err := base64.StdEncoding.DecodeString(*aux.PathName.Content)
+			if err != nil {
+				return err
+			}
+			p.Name = to.Ptr(string(name))
+		} else {
+			p.Name = aux.PathName.Content
+		}
+	}
+	return nil
 }